backend/requirements.txt
anyio==4.12.0
arrow==1.4.0
asgiref==3.10.0
attrs==25.4.0
Automat==25.4.16
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
cssselect==1.3.0
defusedxml==0.7.1
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
elastic-transport==8.17.1
elasticsearch==8.19.1
elasticsearch-dsl==8.18.0
fastjsonschema==2.21.2
filelock==3.20.0
fqdn==1.5.1
graphql-core==3.2.7
h11==0.16.0
harfile==0.4.0
httpcore==1.0.9
httpie==3.2.4
httpx==0.28.1
hyperlink==21.0.0
hypothesis==6.148.7
hypothesis-graphql==0.11.1
hypothesis-jsonschema==0.23.1
idna==3.11
incremental==24.7.2
iniconfig==2.3.0
isoduration==20.11.0
itemadapter==0.12.2
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
junit-xml==1.9
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mistune==3.2.0
multidict==6.7.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.3
packaging==25.0
pandocfilters==1.5.1
parsel==1.10.0
platformdirs==4.5.1
pluggy==1.6.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyDispatcher==2.0.7
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyrate-limiter==3.9.0
PySocks==1.7.1
pytest==9.0.2
pytest-subtests==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
queuelib==1.8.0
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==14.2.0
rpds-py==0.30.0
schemathesis==4.7.6
scipy==1.16.3
Scrapy==2.13.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.1
sqlparse==0.5.3
starlette==0.50.0
starlette-testclient==0.4.1
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tldextract==5.3.0
tornado==6.5.4
traitlets==5.14.3
Twisted==25.5.0
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
w3lib==2.3.1
webcolors==25.10.0
webencodings==0.5.1
Werkzeug==3.1.4
zope.interface==8.0.1
